Tarefa1_2022li1g034

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2022/23.

Documentation

mapaValido :: Mapa -> Bool

Função Principal

A função mapaValido, que recebe um Mapa e retorna Bool, é composta pela conjunção de funções auxiliares, que a permitem validar o mapa recebido. Basicamente, se todas as funções auxiliares validarem o mapa recebido, este é válido e pode ser usado no jogo.

A função é assim definida por :

mapaValido (Mapa l []) = True  
mapaValido (Mapa l t) = obsImproprios (Mapa l t) && 
                        veloRios (Mapa l t) && 
                        troncosSeguidos (Mapa l t) &&
                        carrosSeguidos (Mapa l t) &&
                        elemNenhum (Mapa l t) &&
                        larguraObs (Mapa l t) &&
                        terrSeguidos (Mapa l t)

obsImproprios :: Mapa -> Bool

Função Auxiliar

obsImproprios é uma função auxiliar que verifica se existem obstáculos em terrenos impróprios, e.g. troncos em estradas ou relvas, árvores em rios ou estradas, etc. Caso exista é um Mapa inválido.

veloRios :: Mapa -> Bool

Função Auxiliar

veloRios é uma função auxiliar que verifica se os rios contíguos têm velociades opostas. Se rios contíguos tiverem velociades paralelas (ambas positivas ou ambas negativas) o Mapa é inválido.

troncosSeguidos :: Mapa -> Bool

Função Auxiliar

troncosSeguidos é uma função auxiliar que verifica se os troncos têm, no máximo, 5 unidades de comprimento. Caso haja mais de 5 troncos contíguos numa linha circular, o Mapa é inválido.

troncosSeguidos' :: [Obstaculo] -> Int -> Bool

Função Auxiliar da Auxiliar

troncosSeguidos' é uma função auxiliar da função troncosSeguidos.

carrosSeguidos :: Mapa -> Bool

Função Auxiliar

carrosSeguidos é uma função auxiliar que verifica se os carros têm, no máximo, 3 unidades de comprimento. Caso haja mais de 3 carros contíguos numa linha, o Mapa é inválido.

carrosSeguidos' :: [Obstaculo] -> Int -> Bool

Função Auxiliar da Auxiliar

carrosSeguidos' é uma função auxiliar da função carrosSeguidos.

elemNenhum :: Mapa -> Bool

Função Auxiliar

elemNenhum é uma função auxiliar que verifica a existência de, pelo menos um, "obstáculo" Nenhum. Ou seja, cada linha deve conter um espaço livre. Se isso não acontecer, o Mapa é inválido.

larguraObs :: Mapa -> Bool

Função Auxiliar

larguraObs é uma função auxiliar da função que verifica se o comprimento da lista de obstáculos corresponde exatamente à largura do Mapa. Caso o comprimento da lista seja maior ou menor, o Mapa é inválido.

terrSeguidos :: Mapa -> Bool

Função Auxiliar

terrSeguidos é uma função auxiliar que verifica se existem mais de 4 rios contíguos ou mais de 5 estradas contíguas ou 5 relvas contíguas. Caso exista, o Mapa é inválido.

terrSeguidos' :: [(Terreno, [Obstaculo])] -> Int -> Bool

Função Auxiliar da Auxiliar

terrSeguidos' é uma função auxiliar da função terrSeguidos.

validaPassagem :: Mapa -> Bool

Função Auxiliar

terrSeguidos' é uma função auxiliar que verifica se existem passagens possíveis de uma linha para outra ao longo do mapa.