Tarefa1_2022li1g073

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2022/23.

Documentation

mapaValido :: Mapa -> Bool

A função "mapaValido", através da concordância entre as 10 funções auxiliares que a compõem, verifica se um dado mapa é válido, isto é, não viola nenhuma das restrições impostas pelas funções auxiliares.

mapaValido :: Mapa -> Bool
mapaValido (Mapa l t) = mapaValido1 (Mapa l t) 
                       && mapaValido2 (Mapa l t)
                       && mapaValido3 (Mapa l t)
                       && mapaValido3excecao (Mapa l t)
                       && mapaValido4 (Mapa l t)
                       && mapaValido4excecao (Mapa l t)
                       && mapaValido5 (Mapa l t)
                       && mapaValido5especial (Mapa l t)
                       && mapaValido6 (Mapa l t)
                       && mapaValido7 (Mapa l t)
                       && mapaValido8 l t

mapaValido1 :: Mapa -> Bool

A função "mapaValido1" garante que não existem obstáculos em terrenos impróprios, e.g. troncos em estradas ou relvas, árvores em rios ou estradas, etc.

mapaValido2 :: Mapa -> Bool

A função "mapaValido2" garante que rios contíguos têm direções opostas.

mapaValido3 :: Mapa -> Bool

A função "mapaValido3" garante que troncos têm, no máximo, 5 unidades de comprimento.

mapaValido3aux :: [Obstaculo] -> Int -> Bool

A função "mapaValido3aux" é uma função necessária para alcançar o propósito da função "mapaValido3".

mapaValido3excecao :: Mapa -> Bool

Tal como o nome indica, a função "mapaValido3excecao" não é mais do que uma exceção, exceção essa que tem em conta não só a regra imposta pela função "mapaValido3", como também o metódo funcional do jogo a um nível futuro, garantindo assim que não ocorreram erros inesperados.

mapaValido4 :: Mapa -> Bool

A função "mapaValido4" garante que carros têm, no máximo, 3 unidades de comprimento.

mapaValido4aux :: [Obstaculo] -> Int -> Bool

A função "mapaValido4aux" é uma função necessária para alcançar o propósito da função "mapaValido4".

mapaValido4excecao :: Mapa -> Bool

À semelhança da "mapaValido3excecao", a função "mapaValido4excecao tem também em conta fases futuras da construção do nosso jogo, revela-se por isso necessária quando tendo em conta a regra imposta pela função "mapaValido4".

mapaValido5 :: Mapa -> Bool

A função "mapaValido5" garante que em qualquer linha existe, no mínimo, um "obstáculo" Nenhum. Ou seja, uma linha não pode ser composta exclusivamente por obstáculos, precisando de haver pelo menos um espaço livre.

mapaValido5especial :: Mapa -> Bool

A função "mapaValido5especial" foi criada com base no raciocínio conceptual abordado na função "mapaValido5", assim como tem de existir no mínimo um "obstáculo" Nenhum para a existência de espaços livres, tem também de existir, pelo menos um "obstáculo" Tronco quando nos encontramos no Terreno Rio para que o jogador consiga atravessar esse mesmo rio.

mapaValido6 :: Mapa -> Bool

A função "mapaValido6" garante que o comprimento da lista de obstáculos de cada linha corresponde exatamente à largura do mapa.

mapaValido7 :: Mapa -> Bool

A função "mapaValido7" garante que contiguamente, não devem existir mais do que 4 rios, nem 5 estradas ou relvas.

mapaValido8 :: Largura -> [(Terreno, [Obstaculo])] -> Bool

A função "mapaValido8" limita o número de obstáculos numa linha para evitar linhas seguidas com vários obstáculos a obstruir o caminho.