Tarefa2_2022li1g034

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2022/23.

Documentation

estendeMapa :: Mapa -> Int -> Mapa

Função Principal

A função estendeMapa, que recebe um Mapa e retorna Mapa, é composta pela conjunção de funções auxiliares, que deve gerar e adicionar uma linha valida ao topo do mapa recebido. Basicamente, a funcao aceita o seed que e usado para generar o terreno valido. depois usando aquele terreno genera os obstaculos validos. Usando estes dados construe a nova mapa valida. a funcao auxiliar forLoop e usada para generar a lista valida dos obstaculos. A função é assim definida por :

estendeMapa :: Mapa -> Int -> Mapa
estendeMapa (Mapa width []) randomNumber= 
    let 
        velocidadeArr = [1, -1, 2, -2]
        terrenoGenerated = (proximosTerrenosValidos (Mapa width []))!!(mod randomNumber 3)
        finalTerreno = (if(terrenoGenerated == Rio 0) then Rio (velocidadeArr !! (mod randomNumber 4)) else if (terrenoGenerated == Estrada 0) then Estrada (velocidadeArr !! (mod randomNumber 4)) else Relva)
        randomList = geraListaAleatorios randomNumber width
        forLoop :: Int -> Int -> [Obstaculo] -> [Obstaculo]
        forLoop i j accum | i<j = (forLoop (i+1) j (accum ++ [chosenObstacle]))
                          | otherwise = accum
            where
                obstacleOptions = proximosObstaculosValidos width (finalTerreno, accum)
                chosenObstacle = obstacleOptions !! (if (length obstacleOptions) == 2 then mod (randomList !! i) 2 else mod (randomList !! i) 1)
        finalObstaculos = forLoop 0 width []
    in
        (Mapa width [(finalTerreno, finalObstaculos)])
estendeMapa (Mapa width (x:xs)) randomNumber = 
    let 
        velocidadeArr = [1, -1, 2, -2]
        terrenoGenerated = (proximosTerrenosValidos (Mapa width (x:xs))) !! (if length ((proximosTerrenosValidos (Mapa width (x:xs)))) == 3 then mod randomNumber 3 else mod randomNumber 2)
        finalTerreno = (if(terrenoGenerated == Rio 0) then Rio (velocidadeArr !! (mod randomNumber 4)) else if (terrenoGenerated == Estrada 0) then Estrada (velocidadeArr !! (mod randomNumber 4)) else Relva)
        randomList = geraListaAleatorios randomNumber width
        forLoop :: Int -> Int -> [Obstaculo] -> [Obstaculo]
        forLoop i j accum | i<j = (forLoop (i+1) j (accum ++ [chosenObstacle]))
                          | otherwise = accum
            where
                obstacleOptions = proximosObstaculosValidos width (finalTerreno, accum)
                chosenObstacle = obstacleOptions !! (if (length obstacleOptions) == 2 then mod (randomList !! i) 2 else mod (randomList !! i) 1)
        finalObstaculos = forLoop 0 width []
    in
        (Mapa width ([(finalTerreno, finalObstaculos)]++(x:xs)))

geraListaAleatorios :: Int -> Int -> [Int]

Função Auxiliar

geraListaAleatorios gera uma lista de aleatórios de acordo com a semente e o comprimento que pretende.

proximosTerrenosValidos :: Mapa -> [Terreno]

Função Auxiliar

proximosTerrenosValidos é uma função auxiliar que retorna a lista dos proximos terrenos validos (possiveis) para a mapa dada.

proximosObstaculosValidos :: Int -> (Terreno, [Obstaculo]) -> [Obstaculo]

Função Auxiliar

proximosObstaculosValidos é uma função auxiliar que retorna a lista de obstaculos validos para o terreno dado.

obsIniciais :: [Obstaculo] -> [Obstaculo]

Função Auxiliar

obsIniciais é uma função auxiliar que cria uma lista com os obstaculos iniciais iguais.

obsSeguidos :: Int -> Obstaculo -> [Obstaculo] -> Int

Função Auxiliar

obsSeguidos é uma função auxiliar que diz quantos obstaculos seguidos no tem no final da lista.