Tarefa2_2022li1g035

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2022/23.

Documentation

type Comprimento = Int

Aleatoriedade

As funções que se seguem são as responsáveis por atribuir uma aleatoriedade à função estendeMapa:

geraListaAleatorios  :: Semente -> Comprimento -> [Int]
geraListaAleatorios s c= take c $ randoms (mkStdGen s)

nAleatorio :: Semente -> Comprimento -> Int
nAleatorio s c = head(geraListaAleatorios s c)

geradorVelocidades :: Int ->  Int
geradorVelocidades random = (mod random 5)-2

Exemplos de utilização:

>>> estendeMapa (Mapa 3 [(Relva,[Arvore,Arvore,Nenhum])]) 12
Mapa 3 [(Rio 0,[Tronco,Nenhum,Tronco]),(Relva,[Arvore,Arvore,Nenhum])]

type Semente = Int

geraListaAleatorios :: Semente -> Comprimento -> [Int]

nAleatorio :: Semente -> Comprimento -> Int

geradorVelocidades :: Int -> Int

estendeMapa :: Mapa -> Int -> Mapa

estendeMapa

A funçõo estendeMapa deve gerar e adicionar uma nova linha valida ao topo de um dado mapa. Alternativamente, a função poderia ser definida da seguinte forma:

estendeMapa :: Mapa -> Int -> Mapa
estendeMapa (Mapa l linhas) x = Mapa l ((juntaTudo x (Mapa l linhas) ((proximoTerreno x (Mapa l linhas)))): linhas)

Exemplos de utilização:

>>> estendeMapa (Mapa 3 [(Relva,[Arvore,Arvore,Nenhum])]) 12
Mapa 3 [(Rio 0,[Tronco,Nenhum,Tronco]),(Relva,[Arvore,Arvore,Nenhum])]

proximoTerreno :: Int -> Mapa -> (Terreno, [Obstaculo])

justTerreno :: (Terreno, [Obstaculo]) -> Terreno

proximosTerrenosValidos :: Int -> Mapa -> [Terreno]

juntaTudo :: Int -> Mapa -> (Terreno, [Obstaculo]) -> (Terreno, [Obstaculo])

A função juntaTudo deve juntar tanto o terreno gerado como a lista de obstáculos que tambem foi gerada e assim formar uma nova linha. Alternativamente, a função poderia ser definida da seguinte forma:

juntaTudo ::  Int -> Mapa -> (Terreno,[Obstaculo]) -> (Terreno,[Obstaculo])
juntaTudo x (Mapa l linhas) (t, obs) = (t,proximoObs x (Mapa l linhas) (t, obs))

Exemplos de utilização:

>>> juntaTudo 12 (Mapa 3 [(Relva,[Arvore,Arvore,Nenhum])]) (Rio 0,[])
(Rio 0,[Tronco,Nenhum,Tronco])

proximoObs :: Int -> Mapa -> (Terreno, [Obstaculo]) -> [Obstaculo]

proximosObstaculosValidos :: Int -> (Terreno, [Obstaculo]) -> [Obstaculo]

proximosObsAUX :: Int -> (Terreno, [Obstaculo]) -> [Obstaculo]