Tarefa2_2022li1g073

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2022/23.

Documentation

estendeMapa :: Mapa -> Int -> Mapa

Esta é a função principal que adiciona uma nova linha ao topo do mapa

estendeMapa :: Mapa -> Int -> Mapa
estendeMapa (Mapa l ((ter, obs) : t)) seed = (Mapa l (( geraterreno (Mapa l ((ter, obs) : t)) seed , geralinha l ( geraterreno (Mapa l ((ter, obs) : t)) seed , []) seed) : (ter, obs) : t))

proximosTerrenosValidos :: Mapa -> Int -> [Terreno]

Esta função auxiliar recebe o mapa da função estendeMapa e o número aleatório e indica os possíveis terrenos a acrescentar com velocidades aleatórias

geraterreno :: Mapa -> Int -> Terreno

Esta função auxiliar recebe o mapa da função estendeMapa e o número aleatório e gera um terreno válido indicado na função proximosTerrenosValidos

proximosObstaculosValidos :: Int -> (Terreno, [Obstaculo]) -> [Obstaculo]

Esta função auxiliar recebe a largura do mapa e uma linha do mapa e indica os possíveis obstáculos a acrescentar de acordo com o terreno da linha e a sua largura

geraobstaculo :: Int -> (Terreno, [Obstaculo]) -> Int -> Obstaculo

Esta função auxiliar recebe a largura do mapa, uma linha do mapa e um número aleatório e gera um obstáculo indicado na função proximosObstaculosValidos

geralinha :: Int -> (Terreno, [Obstaculo]) -> Int -> [Obstaculo]

Esta função auxiliar recebe a largura do mapa, uma linha do mapa e um número aleatório e gera a linha de obstáculos respeitando a largura do Mapa

aleatorios :: Int -> Int -> [Int]

Esta função auxiliar gera um número aleatório de 0 a 2 que corresponde à posição de um terreno na lista que a função proximosTerrenosValido devolve

aleatorios1 :: Int -> Int -> [Int]

Esta função auxiliar gera um número aleatório de 0 a 1 que corresponde à posição de um terreno na lista que a função proximosTerrenosValido devolve e à posição de um obstáculo na lista que a função proximosObstaculosValidos devolve

aleatoriosv :: Int -> Int -> [Int]

Esta função auxiliar gera uma velocidade aleatória, diferente de 0 para os terrenos Rio e Estrada