Tarefa3_2022li1g034

Description

Módulo para a realização da Tarefa 3 do projeto de LI1 em 2022/23.

Documentation

animaJogo :: Jogo -> Jogada -> Jogo

Função Principal

animaJogo é a função principal pedida nesta Tarefa e pretende animar o Jogo, ou seja, o Jogador e o Mapa, consoante a respetiva Jogada e lista de Obstaculo apresentados no mapa.

Constituída por funções auxiliares, a função principal é definida por:

animaJogo :: Jogo -> Jogada -> Jogo 
animaJogo (Jogo (Jogador (x,y)) (Mapa l (h:t))) Parado = Jogo jogador1 (Mapa l (movObs (h:t) (x,y) 0 0))
                                                        where jogador1 = (paradoTronco (Jogo (Jogador (x,y)) (Mapa l (h:t))) 0)

animaJogo (Jogo (Jogador (x,y)) (Mapa l (h:t))) (Move Direita) | l <= (x+1) = Jogo jogador1 (Mapa l (movObs (h:t) (x,y) 0 0))
                                                               | posArvore (h:t) (x+1,y) 0 = Jogo (Jogador (x,y)) (Mapa l (movObs (h:t) (x,y) 0 0))  
                                                               | otherwise = Jogo jogador2 (Mapa l (movObs (h:t) (x+1,y) 0 0))
                                                                 where jogador1 = (paradoTronco (Jogo (Jogador (x,y)) (Mapa l (h:t))) 0)  
                                                                       jogador2 = (paradoTronco (Jogo (Jogador (x+1,y)) (Mapa l (h:t))) 0)

animaJogo (Jogo (Jogador (x,y)) (Mapa l (h:t))) (Move Esquerda) | x<=0 = Jogo jogador1 (Mapa l (movObs (h:t) (x,y) 0 0))
                                                                | posArvore (h:t) (x-1,y) 0 = Jogo (Jogador (x,y)) (Mapa l (movObs (h:t) (x-1,y) 0 0)) 
                                                                | otherwise = Jogo jogador2 (Mapa l (movObs (h:t) (x-1,y) 0 0))
                                                                  where jogador1 = (paradoTronco (Jogo (Jogador (x,y)) (Mapa l (h:t))) 0)
                                                                        jogador2 = (paradoTronco (Jogo (Jogador (x-1,y)) (Mapa l (h:t))) 0)

animaJogo (Jogo (Jogador (x,y)) (Mapa l (h:t))) (Move Baixo) | y>=(length (h:t))-1= Jogo (Jogador (x,y)) (Mapa l (movObs (h:t) (x,y) 0 0))
                                                             | posArvore (h:t) (x,y+1) 0 = Jogo jogador1 (Mapa l (movObs (h:t) (x,y) 0 0)) 
                                                             | otherwise = Jogo (Jogador (x,y+1)) (Mapa l (movObs (h:t) (x,y) 0 0))
                                                              where jogador1 = (paradoTronco (Jogo (Jogador (x,y)) (Mapa l (h:t))) 0)

animaJogo (Jogo (Jogador (x,y)) (Mapa l (h:t))) (Move Cima) | y<=0 = Jogo (Jogador (x,y)) (Mapa l (movObs(h:t) (x,y) 0 0))
                                                            | posArvore (h:t) (x,y-1) 0 = Jogo jogador1 (Mapa l (movObs (h:t) (x,y) 0 0)) 
                                                            | otherwise = Jogo (Jogador (x,y-1)) (Mapa l (movObs (h:t) (x,y) 0 0))
                                                             where jogador1 = (paradoTronco (Jogo (Jogador (x,y)) (Mapa l (h:t))) 0)
                                                              
animaJogo (Jogo (Jogador (x,y)) (Mapa l [])) _ = Jogo (Jogador (x,y)) (Mapa l [])

movObs :: [(Terreno, [Obstaculo])] -> (Int, Int) -> Int -> Int -> [(Terreno, [Obstaculo])]

Função Auxiliar

movObs é uma função auxiliar que gere o movimento dos Obstáculos consoante a velocidade do terreno, seja ela positiva ou negativa, ou inexistente no caso da Relva.

paradoTronco :: Jogo -> Int -> Jogador

Função Auxiliar

paradoTronco é uma função auxiliar que permite o movimento do jogador enquanto este se encontra Parado na mesma localização de um tronco, ou seja, segue o movimento do tronco, dependendo da velocidade do terreno e posição do jogador. Também foi utilizado para movimentos laterais, para a soma de vetores da velocidade e jogada, e para movimentos verticais, caso se encontre num rio e queira se movimentar para a mesma posição de uma árvore.

posArvore :: [(Terreno, [Obstaculo])] -> (Int, Int) -> Int -> Bool

Função Auxiliar

posArvore é uma função auxiliar que verifica se a posição que o jogador se encontra é igual à posição duma Arvore.

atropelado :: (Terreno, [Obstaculo]) -> Int -> Bool

Função Auxiliar

atropelado é uma função auxiliar que verifica se a posição pedida apresenta um Carro.

animaJogador :: Jogo -> Jogada -> Jogo

Função Principal

animaJogador é uma função derivada da animaJogo,em que anima apenas os movimentos do Jogador.