Tarefa3_2022li1g035

Description

Módulo para a realização da Tarefa 3 do projeto de LI1 em 2022/23.

Documentation

animaJogo :: Jogo -> Jogada -> Jogo

animaJogo

A funçõo animaJogo é a função principal que automatiza todos os elementos do mapa, sendo eles superfícies, osbtaculos ou o jogador.

Corpo da função:

animaJogo :: Jogo -> Jogada -> Jogo
animaJogo (Jogo (Jogador (x,y)) (Mapa l ((terreno,obs):t))) j = 
    juntaTudo (moveTerreno ((terreno, obs):t)) (jogadaValida (Jogador (x,y)) j (Mapa l ((terreno,obs):t))) l

juntaTudo :: [(Terreno, [Obstaculo])] -> Jogador -> Int -> Jogo

juntaTudo

A funcao juntaTudo recebe os argumentos das funcoes auxiliares moveTerreno e jogadaValida e devolve um argumento do tipo Jogo.

Corpo da função:

juntaTudo :: [(Terreno, [Obstaculo])] -> Jogador -> Int -> Jogo 
juntaTudo ((terreno,obs):t) (Jogador (x,y)) l = (Jogo (Jogador (x,y)) (Mapa l ((terreno,obs):t)))

Exemplos de utilização:

>>> juntaTudo [(Rio 1, [Nenhum, Tronco, Nenhum])] (Jogador (0,0)) 3
Jogo (Jogador (0,0)) (Mapa 3 [(Rio 1,[Nenhum,Tronco,Nenhum])]) 

moveTerreno :: [(Terreno, [Obstaculo])] -> Jogador -> Jogada -> [(Terreno, [Obstaculo])]

moveTerreno

A funcao moveTerreno que faz os obstaculos moverem-se com a velocidade v dos terrenos (Rio ou Estrada)

Corpo da função:

moveTerreno :: [(Terreno, [Obstaculo])] -> [(Terreno, [Obstaculo])]
moveTerreno [] = []
moveTerreno ((terreno, obs):t)
    | isRio terreno = (terreno, transportador obs (getVelocidade terreno) 0) : moveTerreno t
    | isEstrada terreno = (terreno, transportador obs (getVelocidade terreno) 0) : moveTerreno t

Exemplos de utilização:

>>> moveTerreno [(Rio 1, [Nenhum, Tronco, Nenhum])]
[(Rio 1,[Nenhum,Nenhum,Tronco])] 

transportadorEstrada :: [(Terreno, [Obstaculo])] -> Int -> Int -> Jogador -> Jogada -> [Obstaculo]

getVelocidade :: Terreno -> Int

getVelocidade

A funcao getVelocidade retira o valor da velocidade de um determinado terreno == Corpo da função:

getVelocidade :: Terreno -> Int
getVelocidade (Rio v) = v
getVelocidade (Estrada v) = v

Exemplos de utilização:

>>> getVelocidade (Rio 1)
1 

transportador :: [Obstaculo] -> Int -> Int -> [Obstaculo]

transportador

A funcao transportador transporta os elementos de uma ponta da lista para outra conforme a velocidade do terreno. == Corpo da função:

transportador :: [Obstaculo] -> Int -> Int -> [Obstaculo]
transportador (x:xs) v ac 
    | v == ac = (x:xs)
    | v > 0 = transportador (last xs : init (x:xs)) v (ac+1)
    | otherwise = transportador (tail (x:xs) ++ [x]) v (ac-1)

Exemplos de utilização:

>>> transportador [Tronco, Tronco, Nenhum] 1 0
[Nenhum,Tronco,Tronco] 

moveTerrenoAux :: Mapa -> [[Obstaculo]]

jogadaValida :: Jogador -> Jogada -> Mapa -> Jogador

jogadaValida

A funcao jogadaValida verifica se a jogada pode ser realizada e realiza-a.

jogadaValida :: Jogador -> Jogada -> Mapa -> Jogador
jogadaValida (Jogador (x,y)) j (Mapa l ((terreno,obs):t))
    | myFst (moveJogador (Jogador (x,y)) j (Mapa l ((terreno,obs):t))) > l = Jogador (x,y)
    | (mySnd (moveJogador (Jogador (x,y)) j (Mapa l ((terreno,obs):t))) || (mySnd (moveJogador (Jogador (x,y)) j (Mapa l ((terreno,obs):t)))) (length ((terreno,obs):t))) == True = Jogador (x,y)
    | otherwise = moveJogador (Jogador (x,y)) j (Mapa l ((terreno,obs):t))

Exemplos de utilização:

>>> jogadaValida (Jogador (2,1)) (Move Cima) (Mapa 3 [(Rio 1,[Nenhum,Nenhum,Tronco]), (Relva, [Arvore, Nenhum, Nenhum])])
Jogador (2,0) 

moveJogador :: Jogador -> Jogada -> Mapa -> Jogador

moveJogador

A funcao moveJogador faz o jogador mover-se com base nas suas coordenadas e no terreno em que ele se encontra caso seja permitido. Esta funcao serve como auxiliar para a funcao jogadaValida. == Corpo da função:

moveJogador :: Jogador -> Jogada -> Mapa -> Jogador
moveJogador (Jogador (x,y)) Parado (Mapa l p) 
    | ( snd ( p !! y ) !! (x) ) == Tronco = Jogador (x + getVelocidade ( fst (p !! y ) ), y) 
    | otherwise = Jogador (x,y)
moveJogador (Jogador (x,y)) (Move Cima) (Mapa l p) = if ( snd ( p !! (y-1)) !! (x) ) == Arvore then Jogador (x,y) else Jogador (x, y-1)
moveJogador (Jogador (x,y)) (Move Baixo) (Mapa l p) = if ( snd ( p !! (y+1) ) !! (x) ) == Arvore then Jogador (x,y) else Jogador (x, y+1) 
moveJogador (Jogador (x,y)) (Move Esquerda) (Mapa l p) = if ( snd ( p !! (y) ) !! (x-1) ) == Arvore then Jogador (x,y) else Jogador (x-1, y)
moveJogador (Jogador (x,y)) (Move Direita) (Mapa l p) = if ( snd ( p !! (y+1) ) !! (x) ) == Arvore then Jogador (x,y) else Jogador (x+1, y)

Exemplos de utilização:

>>> getVelocidade (Rio 1)
1
== Funcoes auxiliares
@
myFst :: Jogador -> Int 
myFst (Jogador (x,y)) = x
mySnd :: Jogador -> Int 
mySnd (Jogador (x,y)) = y
@

myFst :: Jogador -> Int

mySnd :: Jogador -> Int

isRelva :: Terreno -> Bool