Tarefa3_2022li1g073

Description

Módulo para a realização da Tarefa 3 do projeto de LI1 em 2022/23.

Documentation

animaJogo :: Jogo -> Jogada -> Jogo

A função "animaJogo" implementa a etapa relativa à movimentação do personagem e dos obstáculos. Assim sendo, fazendo uso de algumas funções auxiliares, movimenta os obstáculos (de acordo com a velocidade do terreno em que se encontram), e o personagem, de acordo com a jogada dada.

animaJogo :: Jogo -> Jogada -> Jogo
animaJogo (Jogo (Jogador (x,y)) (Mapa l t)) (direcao) = (Jogo player novomapa)
                                  where novomapa = movelinha (Mapa l t) 
                                        posicaonova = move l (Jogador (x,y)) (direcao)
                                        player = posicaorio l (posicaonova) linh1 (direcao) 
                                        linh1 = (!!) t (abs y)

movelinha :: Mapa -> Mapa

A função "movelinha" é o motor do "novo" mapa que será criado pela movimentação dos obstáculos, assim, garante que uma estrada ou rio com velocidade v, os obstáculos devem mover-se |v| unidades na direção determinada e garante ainda que ao deslocar os obstáculos de uma linha, assim que desaparecerem por um dos lados do mapa, estes reapareçam no lado oposto dessa mesma linha.

movelinhaaux :: [(Terreno, [Obstaculo])] -> [(Terreno, [Obstaculo])]

A função "movelinhaaux" auxilia a função "movelinha" a cumprir o seu propósito.

deslocaDireita :: Int -> [a] -> [a]

deslocaEsq :: Int -> [a] -> [a]

As funções "deslocaDireita" e "deslocaEsq" são essencias para o funcionamento da função "movelinhaaux" e, consequentemente, da "movelinha" e participam na deslocação dos obstáculos numa dada linha.

move :: Jogo -> Jogada -> Jogador

A função "move" movimenta o jogador, isto é, altera a posição do jogador de acordo com a jogada que lhe é dada.

foramapa1 :: Largura -> Jogador -> [(Terreno, [Obstaculo])] -> Bool

selectT :: [(Terreno, [Obstaculo])] -> Int -> (Terreno, [Obstaculo])

busca :: Jogo -> (Terreno, Obstaculo)

posicaorio :: Largura -> Jogador -> (Terreno, [Obstaculo]) -> Jogada -> Jogador

A função "posicaorio" garante que se o personagem se encontrar em cima de um tronco, o jogador acompanha o movimento tronco e ainda, que o jogador não pode estar na mesma posição que uma árvore, um obstáculo.

xTronco :: Obstaculo -> [Obstaculo] -> [Int]

A função "xTronco" dá uma lista com o índice das abcissas de um dado obstáculo especificado em uma dada linha, por exemplo na função "posicaorio" é especificado o obstáculo Tronco nos terrenos de Rio e Árvore nos terrenos de Relva.

xTroncoaux :: Obstaculo -> [Obstaculo] -> Int -> [Int]

A função "xTroncoaux" é uma função auxiliar essencial para o bom funcionamento da função "xTronco".