Tarefa4_2022li1g034

Description

Módulo para a realização da Tarefa 4 do projeto de LI1 em 2022/23.

Documentation

jogoTerminou :: Jogo -> Bool

Função Principal

jogoTerminou é a função principal pedida nesta Tarefa que indica se o jogador perdeu o jogo, retornando assim um True, caso o personangem se encontre fora do mapa, na água, ou atropelado (na mesma posição de um carro).

Constituída por funções auxiliares, a função principal é definida por:

jogoTerminou (Jogo (Jogador (x,y)) (Mapa l [])) = True
jogoTerminou (Jogo (Jogador (x,y)) (Mapa l t)) = (x < 0) ||
                                                 (x >= l) || 
                                                 (y < 0) || 
                                                 (y >= length t) || 
                                                 (naAgua t (x,y) 0) ||  
                                                 (atropelado t (x,y) 0) || 
                                                 (emArvore t (x,y) 0)

naAgua :: [(Terreno, [Obstaculo])] -> (Int, Int) -> Int -> Bool

Função Auxiliar

naAgua é uma função auxiliar que verifica se a posição que o jogador se encontra é dentro da água do Rio, ou seja, na mesma posição do obstáculo Nenhum.

atropelado :: [(Terreno, [Obstaculo])] -> (Int, Int) -> Int -> Bool

Função Auxiliar

atropelado é uma função auxiliar que verifica se a posição que o jogador se encontra é a mesma posição do obstáculo Carro, ou seja, verifica se foi atropelado.

emArvore :: [(Terreno, [Obstaculo])] -> (Int, Int) -> Int -> Bool

Função Auxiliar

emArvore é uma função auxiliar que verifica se a posição que o jogador se encontra é igual à posição duma Arvore.