Tarefa4_2022li1g073

Description

Módulo para a realização da Tarefa 4 do projeto de LI1 em 2022/23.

Documentation

jogoTerminou :: Jogo -> Bool

A função "jogoTerminou" como o jogo indica irá traduzir aquilo que seria um "Game Over", no nosso jogo, indica portanto se o jogador perdeu o jogo, onde True significa que sim. Para isso deve testar se o jogador se encontra fora do mapa, na água, ou “debaixo” de um carro, (ou seja, na mesma posição de um carro).

jogoTerminou :: Jogo -> Bool  
jogoTerminou (Jogo (Jogador (x,y)) (Mapa l t)) = morreste (Jogador (x,y)) t 
                                             || foramapa l (Jogador (x,y)) t 

morreste :: Jogador -> [(Terreno, [Obstaculo])] -> Bool

morreste1 :: Jogador -> (Terreno, [Obstaculo]) -> Bool

A função "morreste1" é uma função auxiliar necessário para que a função "morreste" atinja o seu propósito.

xNenhum4 :: Obstaculo -> [Obstaculo] -> [Int]

A função "xNenhum4" dá uma lista com o índice das abcissas de um dado obstáculo especificado em uma dada linha, por exemplo na função "morreste1" é especificado o obstáculo Nenhum nos terrenos de Rio e Carro nos terrenos de Estrada. Consequentemente, a função "morreste" não funcionaria sem a "xNenhum4".

xNenhum4aux :: Obstaculo -> [Obstaculo] -> Int -> [Int]

foramapa :: Largura -> Jogador -> [(Terreno, [Obstaculo])] -> Bool

A função "foramapa" vai testar se o jogador se encontra fora do mapa, se sim então isto quererá dizer que o jogador perdeu, onde True significa que sim.

atropelado :: Jogador -> Mapa -> Bool

atropeladoaux :: Jogador -> (Terreno, [Obstaculo]) -> Bool

testa :: Int -> Int -> [Int] -> Bool

move4 :: Jogador -> Jogada -> Jogador

movelinha4 :: Mapa -> Mapa

movelinhaaux4 :: [(Terreno, [Obstaculo])] -> [(Terreno, [Obstaculo])]

deslocaEsq :: Int -> [a] -> [a]

deslocaDireita :: Int -> [a] -> [a]