Tarefa5_2022li1g034

Description

Módulo para a realização da Tarefa 5 do projeto de LI1 em 2022/23.

Documentation

deslizaJogo :: Jogo -> Jogo

Função principal

deslizaJogo é uma função principal da Tarefa 5 que "desliza" o jogo uma linha, ou seja, cria um novo mapa, retirando a linha final de e acrescentando uma ao topo da lista de (Terreno, [Obstaculo]).

headMapa :: Mapa -> (Terreno, [Obstaculo])

Função auxiliar

headMapa é uma função auxiliar que recebe um mapa e entrega a cabeça da lista de (Terreno, [Obstaculo]).

deslizaJogoValido :: Jogo -> Int -> Jogo

Função auxiliar

deslizaJogoValido é uma função auxiliar que recebe um jogo e um inteiro, criando o mecanismo de deslize do mapa em deslizaJogo, tendo em conta se o mapa é valido e possível de ultrapassar, com ajuda das funções anteriormente feitas estendeMapa, mapaValido e validaPassagem.