Tarefa5_2022li1g073

Description

Módulo para a realização da Tarefa 5 do projeto de LI1 em 2022/23.

Documentation

deslizaJogo :: Int -> Jogo -> Jogo

Esta é a função principal que remove a última linha do mapa e adiciona uma nova linha válida no topo do mapa

deslizaJogo :: Int -> Jogo -> Jogo 
deslizaJogo r (Jogo (Jogador (x,y)) (Mapa l t)) = (Jogo (Jogador (x, y-1)) (nextmapa r (Mapa l tira)))
                                   where tira = init t

nextmapa :: Int -> Mapa -> Mapa

Esta função usa a função "estendeMap" da Tarefa2 e na função "mapaValido" da Tarefa1. Esta faz correr a "estendeMapa" até gerar uma nova linha que é depois validada pelo "mapavalido", se for válida, é adicionada ao mapa, se não for, repete a função com um valor Int diferente para gerar uma linha nova aleatória