Tarefa6_2022li1g073

Description

Módulo para a realização da Tarefa 6 do projeto de LI1 em 2022/23.

Documentation

data Opcao

Estas serão as opções disponíveis no menu inicial do jogo.

Constructors

Jogar

Sair

data Personagens

Constructors

Galinha

Ronaldo

Vaca

Voltar

data Menu

Estas serão todas as telas que podem ser mostradas no nosso jogo.

Constructors

Opcoes Opcao

EscolhaJogador Personagens

ModoJogo Personagens

GameOver Personagens

type Images = [Picture]

type Time = Float

type Mundo = (Menu, Mundojogo, Time)

type Mundojogo = (Jogo, Jogada, Images, Float)

drawOption :: String -> Picture

"drawOption" é uma função que recebe uma String e retorna uma imagem. Essa imagem é um Texto com o conteúdo da String que foi passada como parâmetro.

janela :: Display

"janela" é uma constante que representa uma janela de exibição para o jogo. Ela tem o nome CrossyRoad, mede 900x900 pixels na posição (500, 100). Esta é uma das funções principais do jogo.

desenharmapa :: Mapa -> [Picture] -> Picture

"desenharmapa" é uma função que recebe um mapa e uma lista de imagens e retorna uma imagem. A imagem é composta pelo resultado da função "escolhelinha", que é aplicada ao mapa e à lista de imagens.

desenharJogador1 :: Jogador -> Largura -> [Picture] -> Picture

"desenharJogador" é uma função que recebe um jogador, a largura do mapa e uma imagem e retorna uma imagem. Essa imagem é a imagem do jogador "translated" para a posição correta no mapa de acordo com as coordenadas x e y do jogador. A largura do mapa é usada para calcular a posição correta da imagem do jogador na tela.

desenharJogador2 :: Jogador -> Largura -> [Picture] -> Picture

desenharJogador3 :: Jogador -> Largura -> [Picture] -> Picture

desenharMundo :: Mundo -> Picture

"desenharMundo" é uma função que recebe um mundo, que é um tipo composto por um estado, um par de um jogo e imagens, e um número. Ela retorna uma imagem que representa o mundo de acordo com o estado. Se o estado for GameOver, a imagem exibida será um fundo e um texto informando o número de pontos e segundos sobrevividos pelo jogador. Se o estado for Opções Jogar, a imagem exibida será um fundo com as opções Jogar e Sair. Se o estado for Opções Sair, a imagem exibida será o mesmo que no estado anterior, mas a opção Sair ficará em vermelho. Se o estado for ModoJogo, a imagem exibida será o resultado da aplicação da função desenharmapa ao mapa e à lista de imagens, com a imagem do jogador sobreposta. Esta é uma das funções principais do jogo.

escolhelinha :: Largura -> [(Terreno, [Obstaculo])] -> Int -> Float -> Int -> [Picture] -> [Picture]

"escolhelinha" é uma função recursiva que recebe a largura do mapa, uma lista de tuplas contendo um terreno e uma lista de obstáculos, um índice, uma coordenada y e outro índice. Ela retorna uma lista de imagens, cada uma representando uma linha do mapa. A função aplica a função "linhagrafica" a cada tupla da lista e adiciona o resultado à lista de imagens. Quando o índice atingir o final da lista, a função retorna a lista de imagens.

linhagrafica :: Largura -> (Terreno, [Obstaculo]) -> Int -> [Picture] -> Picture

"linhagrafica" é uma função que recebe a largura do mapa, uma tupla contendo um terreno e uma lista de obstáculos e um índice, e retorna uma imagem. A imagem é composta pelo resultado da função "linhagrafica1", que é aplicada à tupla, e pelo resultado da função "linhagrafica2", que é aplicada à largura do mapa, à tupla e ao índice.

linhagrafica1 :: (Terreno, [Obstaculo]) -> [Picture] -> Picture

"linhagrafica1" é uma função que recebe uma tupla contendo um terreno e uma lista de obstáculos e uma lista de imagens, e retorna uma imagem. A imagem é composta pelas imagens dos obstáculos presentes na lista de obstáculos, sobrepostas à imagem do terreno. A imagem do terreno é escolhida da lista de imagens de acordo com o tipo de terreno presente na tupla.

linhagrafica2 :: Largura -> (Terreno, [Obstaculo]) -> Int -> [Picture] -> [Picture]

"linhagrafica2" é uma função recursiva que recebe a largura do mapa, uma tupla contendo um terreno e uma lista de obstáculos, um índice e uma lista de imagens, e retorna uma lista de imagens. A lista é composta pelas imagens dos obstáculos presentes na lista de obstáculos, que são posicionadas de acordo com a largura do mapa e o índice. Quando o índice atingir o final da lista de obstáculos, a função retorna a lista de imagens.

desenhaobstaculo :: Velocidade -> Obstaculo -> [Picture] -> Picture

A função "desenhaObstaculo" é uma função que recebe uma velocidade, um obstáculo e uma lista de imagens e retorna uma imagem. A imagem retornada é escolhida da lista de imagens de acordo com o tipo de obstáculo passado como parâmetro. Se o obstáculo for Nenhum, a função retorna a imagem Blank.

eventmove :: Event -> Mundo -> Mundo

A função "eventmove" é uma função que recebe um evento, um mundo e retorna um novo mundo. Ela verifica o tipo do evento e atualiza o estado ou as coordenadas do jogador de acordo. Se o evento for um movimento, ela atualiza as coordenadas do jogador usando a função move e retorna o novo mundo. Se o evento for uma tecla de entrada, ela altera o estado do mundo de acordo. Se o evento for qualquer outro, ela retorna o mundo original. Esta é uma das funções principais do jogo.

fr :: Int

"fr" será o que determinará a "velocidade" de atualização do nosso jogo, isto é, a frequência das dinâmicas do jogo. Esta é uma das funções principais do jogo.

mundoinicial :: [Picture] -> Mundo

"mundoinicial" define-se como o estado incial do jogo. Esta é uma das funções principais do jogo.

aumenta :: Int -> Int -> Int

"aumenta" proporciona aleatoriadade à geração do mapa.

time :: Float -> Mundo -> Mundo

"time" é a função base por trás do sistema de pontos do jogo.

reageTempo :: Float -> Mundo -> Mundo

"reageTempo" é o "motor" do jogo, a função que mantém o jogo a atualizar ao longo do tempo. Esta é uma das funções principais do jogo.